+++ R/utils-globals.R
@@ -21,5 +21,6 @@ utils::globalVariables(c(
   ".imp", ".sat", ".opp", ".high_opp",
   "count.linear.1", "count.linear.05", "count.outlier.sal",
   "sd", "count",
-  ".", "opp_sum", "total_count", "reorder"
+  ".", "opp_sum", "total_count", "reorder",
+  "p.imp", "p.sat", "sig.imp", "sig.sat"
 ))